requirements.txt
APScheduler==3.10.1
cx_Oracle==8.3.0
exchangelib==5.0.1
geopy==2.3.0
impyla==0.18.0
joblib==1.2.0
matplotlib==3.7.1
numpy==1.22.4
openpyxl==3.0.7
pandas==1.5.3
pymysql==1.0.3
python_dateutil==2.8.2
requests==2.28.2
tqdm==4.50.2
